start rewrite tests after a deeper study
